[END-192] Support for eIDAS representative attributes.
It's a crazy idea to think that a foreign IdP will ever be allowed to
tell a SP that person X is representative for person Y. It is the SP's
own choice in deciding of how representation should be handled. 

Furthermore, representation does not belong to authentication. It is
authorization, and a foreign IdP simply can not know about the
authorization model of a SP in another country. 

So. We added this functionality mainly for you people that want to block
assertions that contain representative attributes.